many fixes
there is communication with the grpc server
